+++ packages/editor/js/extensions/libs/typography/compatibility/text-align.js
@@ -3,37 +3,32 @@
 export function textAlignFromWPCompatibility({
 	attributes,
 	blockId,
+	insideBlockInspector = true,
 }: {
 	attributes: Object,
 	blockId: string,
+	insideBlockInspector?: boolean,
 }): Object {
 	let wpAlignAttrId = 'textAlign';
 	if (blockId === 'core/paragraph') {
 		wpAlignAttrId = 'align';
 	}
 
+	// Check block-level style (insideBlockInspector) or global style context
+	const textAlign = insideBlockInspector
+		? attributes[wpAlignAttrId]
+		: attributes?.typography?.textAlign;
+
 	// For detecting the text align changer from block editor controls
 	// we have to validate and make sure the value is correct and should be updated
 	if (
-		attributes[wpAlignAttrId] !== undefined &&
-		attributes?.blockeraTextAlign?.value !== attributes[wpAlignAttrId]
+		textAlign !== undefined &&
+		attributes?.blockeraTextAlign?.value !== textAlign
 	) {
-		switch (blockId) {
-			case 'core/paragraph':
-				if (attributes[wpAlignAttrId] !== undefined) {
-					attributes.blockeraTextAlign = {
-						value: attributes[wpAlignAttrId],
-					};
-				}
-				break;
-
-			default:
-				if (attributes[wpAlignAttrId] !== undefined) {
-					attributes.blockeraTextAlign = {
-						value: attributes?.textAlign,
-					};
-				}
-				break;
+		if (textAlign !== undefined) {
+			attributes.blockeraTextAlign = {
+				value: textAlign,
+			};
 		}
 	}
 
@@ -44,10 +39,12 @@ export function textAlignToWPCompatibility({
 	newValue,
 	ref,
 	blockId,
+	insideBlockInspector = true,
 }: {
 	newValue: Object,
 	ref?: Object,
 	blockId: string,
+	insideBlockInspector?: boolean,
 }): Object {
 	// use correct id for WP data attribute
 	let wpAlignAttrId = 'textAlign';
@@ -60,12 +57,24 @@ export function textAlignToWPCompatibility({
 		'reset' === ref?.current?.action ||
 		['left', 'center', 'right'].indexOf(newValue) === -1
 	) {
-		return {
-			[wpAlignAttrId]: undefined,
-		};
+		return insideBlockInspector
+			? {
+					[wpAlignAttrId]: undefined,
+			  }
+			: {
+					typography: {
+						textAlign: undefined,
+					},
+			  };
 	}
 
-	return {
-		[wpAlignAttrId]: newValue,
-	};
+	return insideBlockInspector
+		? {
+				[wpAlignAttrId]: newValue,
+		  }
+		: {
+				typography: {
+					textAlign: newValue,
+				},
+		  };
 }